RecoMTD: extract TrackSegments into shared TimingTools header
