+++ RecoMTD/TimingTools/BuildFile.xml
@@ -1,0 +1,8 @@
+<use name="DataFormats/Math"/>
+<use name="DataFormats/TrackerRecHit2D"/>
+<use name="FWCore/MessageLogger"/>
+<use name="FWCore/Utilities"/>
+<use name="clhep"/>
+<export>
+  <lib name="1"/>
+</export>
