+++ RecoMTD/TrackExtender/plugins/TrackExtenderWithMTD.cc
@@ -3,6 +3,10 @@
 
 #include <CLHEP/Units/GlobalPhysicalConstants.h>
 
+#include "RecoMTD/TimingTools/interface/MTDHitMatchingInfo.h"
+#include "RecoMTD/TimingTools/interface/TrackSegments.h"
+#include "RecoMTD/TimingTools/interface/TrackTofPidInfo.h"
+
 #include "DataFormats/ForwardDetId/interface/BTLDetId.h"
 #include "DataFormats/ForwardDetId/interface/ETLDetId.h"
 #include "DataFormats/ForwardDetId/interface/MTDChannelIdentifier.h"
@@ -49,303 +53,16 @@ using namespace std;
 using namespace edm;
 using namespace reco;
 
+using mtd::c_cm_ns;
+using mtd::c_inv;
+using mtd::computeTrackTofPidInfo;
+using mtd::MTDHitMatchingInfo;
+using mtd::SigmaTofCalc;
+using mtd::TofCalc;
+using mtd::TrackSegments;
+using mtd::TrackTofPidInfo;
+
 namespace {
-  constexpr float c_cm_ns = geant_units::operators::convertMmToCm(CLHEP::c_light);  // [mm/ns] -> [cm/ns]
-  constexpr float c_inv = 1.0f / c_cm_ns;
-
-  class MTDHitMatchingInfo {
-  public:
-    MTDHitMatchingInfo() {
-      hit = nullptr;
-      estChi2 = std::numeric_limits<float>::max();
-      timeChi2 = std::numeric_limits<float>::max();
-    }
-
-    //Operator used to sort the hits while performing the matching step at the MTD
-    inline bool operator<(const MTDHitMatchingInfo& m2) const {
-      //only for good matching in time use estChi2, otherwise use mostly time compatibility
-      constexpr float chi2_cut = 10.f;
-      constexpr float low_weight = 3.f;
-      constexpr float high_weight = 8.f;
-      if (timeChi2 < chi2_cut && m2.timeChi2 < chi2_cut)
-        return chi2(low_weight) < m2.chi2(low_weight);
-      else
-        return chi2(high_weight) < m2.chi2(high_weight);
-    }
-
-    inline float chi2(float timeWeight = 1.f) const { return estChi2 + timeWeight * timeChi2; }
-
-    const MTDTrackingRecHit* hit;
-    float estChi2;
-    float timeChi2;
-  };
-
-  class TrackSegments {
-  public:
-    TrackSegments() {
-      sigmaTofs_.reserve(30);  // observed upper limit on nSegments
-    };
-
-    inline uint32_t addSegment(float tPath, float tMom2, float sigmaMom) {
-      segmentPathOvc_.emplace_back(tPath * c_inv);
-      segmentMom2_.emplace_back(tMom2);
-      segmentSigmaMom_.emplace_back(sigmaMom);
-      nSegment_++;
-
-      LogTrace("TrackExtenderWithMTD") << "addSegment # " << nSegment_ << " s = " << tPath
-                                       << " p = " << std::sqrt(tMom2) << " sigma_p = " << sigmaMom
-                                       << " sigma_p/p = " << sigmaMom / std::sqrt(tMom2) * 100 << " %";
-
-      return nSegment_;
-    }
-
-    inline float computeTof(float mass_inv2) const {
-      float tof(0.f);
-      for (uint32_t iSeg = 0; iSeg < nSegment_; iSeg++) {
-        float gammasq = 1.f + segmentMom2_[iSeg] * mass_inv2;
-        float beta = std::sqrt(1.f - 1.f / gammasq);
-        tof += segmentPathOvc_[iSeg] / beta;
-
-        LogTrace("TrackExtenderWithMTD") << " TOF Segment # " << iSeg + 1 << " p = " << std::sqrt(segmentMom2_[iSeg])
-                                         << " tof = " << tof;
-
-#ifdef EDM_ML_DEBUG
-        float sigma_tof = segmentPathOvc_[iSeg] * segmentSigmaMom_[iSeg] /
-                          (segmentMom2_[iSeg] * sqrt(segmentMom2_[iSeg] + 1 / mass_inv2) * mass_inv2);
-
-        LogTrace("TrackExtenderWithMTD") << "TOF Segment # " << iSeg + 1 << std::fixed << std::setw(6)
-                                         << " tof segment = " << segmentPathOvc_[iSeg] / beta << std::scientific
-                                         << "+/- " << sigma_tof << std::fixed
-                                         << "(rel. err. = " << sigma_tof / (segmentPathOvc_[iSeg] / beta) * 100
-                                         << " %)";
-#endif
-      }
-
-      return tof;
-    }
-
-    inline float computeSigmaTof(float mass_inv2) {
-      float sigmatof = 0.;
-
-      // remove previously calculated sigmaTofs
-      sigmaTofs_.clear();
-
-      // compute sigma(tof) on each segment first by propagating sigma(p)
-      // also add diagonal terms to sigmatof
-      float sigma = 0.;
-      for (uint32_t iSeg = 0; iSeg < nSegment_; iSeg++) {
-        sigma = segmentPathOvc_[iSeg] * segmentSigmaMom_[iSeg] /
-                (segmentMom2_[iSeg] * sqrt(segmentMom2_[iSeg] + 1 / mass_inv2) * mass_inv2);
-        sigmaTofs_.push_back(sigma);
-
-        sigmatof += sigma * sigma;
-      }
-
-      // compute sigma on sum of tofs assuming full correlation between segments
-      for (uint32_t iSeg = 0; iSeg < nSegment_; iSeg++) {
-        for (uint32_t jSeg = iSeg + 1; jSeg < nSegment_; jSeg++) {
-          sigmatof += 2 * sigmaTofs_[iSeg] * sigmaTofs_[jSeg];
-        }
-      }
-
-      return sqrt(sigmatof);
-    }
-
-    inline uint32_t size() const { return nSegment_; }
-
-    inline uint32_t removeFirstSegment() {
-      if (nSegment_ > 0) {
-        segmentPathOvc_.erase(segmentPathOvc_.begin());
-        segmentMom2_.erase(segmentMom2_.begin());
-        nSegment_--;
-      }
-      return nSegment_;
-    }
-
-    inline std::pair<float, float> getSegmentPathAndMom2(uint32_t iSegment) const {
-      if (iSegment >= nSegment_) {
-        throw cms::Exception("TrackExtenderWithMTD") << "Requesting non existing track segment #" << iSegment;
-      }
-      return std::make_pair(segmentPathOvc_[iSegment], segmentMom2_[iSegment]);
-    }
-
-    uint32_t nSegment_ = 0;
-    std::vector<float> segmentPathOvc_;
-    std::vector<float> segmentMom2_;
-    std::vector<float> segmentSigmaMom_;
-
-    std::vector<float> sigmaTofs_;
-  };
-
-  struct TrackTofPidInfo {
-    float tmtd;
-    float tmtderror;
-    float pathlength;
-
-    float betaerror;
-
-    float dt;
-    float dterror;
-    float dterror2;
-    float dtchi2;
-
-    float dt_best;
-    float dterror_best;
-    float dtchi2_best;
-
-    float gammasq_pi;
-    float beta_pi;
-    float dt_pi;
-    float sigma_dt_pi;
-
-    float gammasq_k;
-    float beta_k;
-    float dt_k;
-    float sigma_dt_k;
-
-    float gammasq_p;
-    float beta_p;
-    float dt_p;
-    float sigma_dt_p;
-
-    float prob_pi;
-    float prob_k;
-    float prob_p;
-  };
-
-  enum class TofCalc { kCost = 1, kSegm = 2, kMixd = 3 };
-  enum class SigmaTofCalc { kCost = 1, kSegm = 2, kMixd = 3 };
-
-  const TrackTofPidInfo computeTrackTofPidInfo(float magp2,
-                                               float length,
-                                               TrackSegments trs,
-                                               float t_mtd,
-                                               float t_mtderr,
-                                               float t_vtx,
-                                               float t_vtx_err,
-                                               bool addPIDError = true,
-                                               TofCalc choice = TofCalc::kCost,
-                                               SigmaTofCalc sigma_choice = SigmaTofCalc::kCost) {
-    constexpr float m_pi = 0.13957018f;
-    constexpr float m_pi_inv2 = 1.0f / m_pi / m_pi;
-    constexpr float m_k = 0.493677f;
-    constexpr float m_k_inv2 = 1.0f / m_k / m_k;
-    constexpr float m_p = 0.9382720813f;
-    constexpr float m_p_inv2 = 1.0f / m_p / m_p;
-
-    TrackTofPidInfo tofpid;
-
-    tofpid.tmtd = t_mtd;
-    tofpid.tmtderror = t_mtderr;
-    tofpid.pathlength = length;
-
-    auto deltat = [&](const float mass_inv2, const float betatmp) {
-      float res(1.f);
-      switch (choice) {
-        case TofCalc::kCost:
-          res = tofpid.pathlength / betatmp * c_inv;
-          break;
-        case TofCalc::kSegm:
-          res = trs.computeTof(mass_inv2);
-          break;
-        case TofCalc::kMixd:
-          res = trs.computeTof(mass_inv2) + tofpid.pathlength / betatmp * c_inv;
-          break;
-      }
-      return res;
-    };
-
-    auto sigmadeltat = [&](const float mass_inv2) {
-      float res(1.f);
-      switch (sigma_choice) {
-        case SigmaTofCalc::kCost:
-          // sigma(t) = sigma(p) * |dt/dp| = sigma(p) * DeltaL/c * m^2 / (p^2 * E)
-          res = tofpid.pathlength * c_inv * trs.segmentSigmaMom_[trs.nSegment_ - 1] /
-                (magp2 * sqrt(magp2 + 1 / mass_inv2) * mass_inv2);
-          break;
-        case SigmaTofCalc::kSegm:
-          res = trs.computeSigmaTof(mass_inv2);
-          break;
-        case SigmaTofCalc::kMixd:
-          float res1 = tofpid.pathlength * c_inv * trs.segmentSigmaMom_[trs.nSegment_ - 1] /
-                       (magp2 * sqrt(magp2 + 1 / mass_inv2) * mass_inv2);
-          float res2 = trs.computeSigmaTof(mass_inv2);
-          res = sqrt(res1 * res1 + res2 * res2 + 2 * res1 * res2);
-      }
-
-      return res;
-    };
-
-    tofpid.gammasq_pi = 1.f + magp2 * m_pi_inv2;
-    tofpid.beta_pi = std::sqrt(1.f - 1.f / tofpid.gammasq_pi);
-    tofpid.dt_pi = deltat(m_pi_inv2, tofpid.beta_pi);
-    tofpid.sigma_dt_pi = sigmadeltat(m_pi_inv2);
-
-    tofpid.gammasq_k = 1.f + magp2 * m_k_inv2;
-    tofpid.beta_k = std::sqrt(1.f - 1.f / tofpid.gammasq_k);
-    tofpid.dt_k = deltat(m_k_inv2, tofpid.beta_k);
-    tofpid.sigma_dt_k = sigmadeltat(m_k_inv2);
-
-    tofpid.gammasq_p = 1.f + magp2 * m_p_inv2;
-    tofpid.beta_p = std::sqrt(1.f - 1.f / tofpid.gammasq_p);
-    tofpid.dt_p = deltat(m_p_inv2, tofpid.beta_p);
-    tofpid.sigma_dt_p = sigmadeltat(m_p_inv2);
-
-    tofpid.dt = tofpid.tmtd - tofpid.dt_pi - t_vtx;  //assume by default the pi hypothesis
-    tofpid.dterror2 = tofpid.tmtderror * tofpid.tmtderror + t_vtx_err * t_vtx_err;
-    tofpid.betaerror = 0.f;
-    if (addPIDError) {
-      tofpid.dterror2 = tofpid.dterror2 + (tofpid.dt_p - tofpid.dt_pi) * (tofpid.dt_p - tofpid.dt_pi);
-      tofpid.betaerror = tofpid.beta_p - tofpid.beta_pi;
-    } else {
-      // only add sigma(TOF) if not considering mass hp. uncertainty
-      tofpid.dterror2 = tofpid.dterror2 + tofpid.sigma_dt_pi * tofpid.sigma_dt_pi;
-    }
-    tofpid.dterror = sqrt(tofpid.dterror2);
-
-    tofpid.dtchi2 = (tofpid.dt * tofpid.dt) / tofpid.dterror2;
-
-    tofpid.dt_best = tofpid.dt;
-    tofpid.dterror_best = tofpid.dterror;
-    tofpid.dtchi2_best = tofpid.dtchi2;
-
-    tofpid.prob_pi = -1.f;
-    tofpid.prob_k = -1.f;
-    tofpid.prob_p = -1.f;
-
-    if (!addPIDError) {
-      //*TODO* deal with heavier nucleons and/or BSM case here?
-      const float dterror2_wo_sigmatof = tofpid.dterror2 - tofpid.sigma_dt_pi * tofpid.sigma_dt_pi;
-      float chi2_pi = tofpid.dtchi2;
-      float chi2_k = (tofpid.tmtd - tofpid.dt_k - t_vtx) * (tofpid.tmtd - tofpid.dt_k - t_vtx) /
-                     (dterror2_wo_sigmatof + tofpid.sigma_dt_k * tofpid.sigma_dt_k);
-      float chi2_p = (tofpid.tmtd - tofpid.dt_p - t_vtx) * (tofpid.tmtd - tofpid.dt_p - t_vtx) /
-                     (dterror2_wo_sigmatof + tofpid.sigma_dt_p * tofpid.sigma_dt_p);
-
-      float rawprob_pi = exp(-0.5f * chi2_pi);
-      float rawprob_k = exp(-0.5f * chi2_k);
-      float rawprob_p = exp(-0.5f * chi2_p);
-      float normprob = 1.f / (rawprob_pi + rawprob_k + rawprob_p);
-
-      tofpid.prob_pi = rawprob_pi * normprob;
-      tofpid.prob_k = rawprob_k * normprob;
-      tofpid.prob_p = rawprob_p * normprob;
-
-      float prob_heavy = 1.f - tofpid.prob_pi;
-      constexpr float heavy_threshold = 0.75f;
-
-      if (prob_heavy > heavy_threshold) {
-        if (chi2_k < chi2_p) {
-          tofpid.dt_best = (tofpid.tmtd - tofpid.dt_k - t_vtx);
-          tofpid.dtchi2_best = chi2_k;
-        } else {
-          tofpid.dt_best = (tofpid.tmtd - tofpid.dt_p - t_vtx);
-          tofpid.dtchi2_best = chi2_p;
-        }
-      }
-    }
-    return tofpid;
-  }
 
   bool getTrajectoryStateClosestToBeamLine(const Trajectory& traj,
                                            const reco::BeamSpot& bs,
